Fixed structure of project, optimized Feed.jsx
