+++ src/Components/PostComponents/EngagementButtons.jsx
@@ -1,0 +1,21 @@
+const EngagementButtons = ({
+  postId,
+  isLiked,
+  likes,
+  commentCount,
+  onLike,
+  onComment
+}) => {
+  return (
+    <div className="engagement-buttons">
+      <button onClick={() => onLike(postId)}>
+        {isLiked ? "🖤" : "❤️"} {likes}
+      </button>
+      <button>
+        💬 {commentCount}
+      </button>
+    </div>
+  );
+};
+
+export default EngagementButtons;
