+++ src/css/Feed.css
@@ -291,6 +291,7 @@ html {
 }
 
 .engagement-buttons button {
+  margin-top: 5px;
   background: none;
   border: none;
   font-size: 16px;
